Rewrite to different split between the providers
